+++ deformo/Loader.h
@@ -5,6 +5,6 @@
 
 namespace loader {
 void ReadTetgenVertexFile(Eigen::MatrixXf& V, const std::string& node_file);
-void ReadTetgenFaceFile(Eigen::MatrixXf& F, const std::string& face_file);
-void ReadTetgenEleFile(Eigen::MatrixXf& T, const std::string& ele_file);
+void ReadTetgenFaceFile(Eigen::MatrixXi& F, const std::string& face_file);
+void ReadTetgenEleFile(Eigen::MatrixXi& T, const std::string& ele_file);
 }  // namespace loader
